refactor: click event 디렉토리 이동
- view에 있던 click event 파일 -> controller로 이동
