+++ src/controller/ConfigureRacingCar.js
@@ -1,6 +1,6 @@
 import { CAUTION_MESSAGE, DOM } from '../utils/constant.js';
-import CarNamesEvent from '../view/CarNamesEvent.js';
-import RacingCountEvent from '../view/RacingCountEvent.js';
+import CarNamesEvent from './CarNamesEvent.js';
+import RacingCountEvent from './RacingCountEvent.js';
 import Game from '../model/Game.js';
 
 export default class ConfigureRacingCar {
